+++ CHANGELOG.md
@@ -8,6 +8,10 @@
 - [#2601](https://github.com/influxdata/kapacitor/pull/2601): Switched to github.com/golang-jwt/jwt for kapacitor's use because of https://nvd.nist.gov/vuln/detail/CVE-2020-26160
 - [#2618](https://github.com/influxdata/kapacitor/pull/2618): Switch task service to use Flux formatter that preserves comments
 
+### Features
+
+- [#2621](https://github.com/influxdata/kapacitor/pull/2621): Add template-id to task list.
+
 ## v1.6.1 [2021-07-22]
 
 ### Features